Create a new endpoint that integrates with a different API:
The API you will integrate with is an API that can give you information on many cryptocurrencies.
The URL you will need is: http://api.coincap.io/v2/assets/bitcoin - the API is free to use and does not require authentication.
Create an endpoint that uses this API to return the current cost of 1 Bitcoin.
Make sure each method is unit test and integration tested as appropriate.